chore: add some initial basic tests for config/creds ini interactions
This commit just adds some early unit tests for the ini_config
module.  A minor refactor was required for some separation of
concerns so that the tests could be written without writing
files to disk.

In subsequent commits the functions in the ini_config will be
refactored further to make them more testable, but this commit
gets the tests in place so that we can hopefully ensure that
we don't break any functionality in the subsequent refactor.

The end goal here is to fix some issues that users reported
where empty tokens don't get updated properly, or leading/trailing
whitespace entered in the CLI input cause corruption to the
files; this will be addressed in later commits once the code
can be tested more easily.